Adding a specific model for the edit page
Updating jquery version

Modified the inline resources view to check if a file is local

+++ lib/presentation/views/vscinlineresources.class.php
@@ -2,7 +2,7 @@
 class vscInlineResources extends vscHtml5View {
 
 	static public function outputScript ($sPath) {
-		if (preg_match('#http[s]?://#', $sPath) == 0) {
+		if (preg_match('#http[s]?://#', $sPath) == 0 && is_file($sPath)) {
 			echo '<script type="text/javascript">' ."\n" . file_get_contents($sPath) . "\n" . '</script>' . "\n";
 		} else {
 			echo '<script type="text/javascript" src="'. $sPath . '"> </script>' . "\n";
@@ -10,7 +10,7 @@ class vscInlineResources extends vscHtml5View {
 	}
 
 	static public function outputStyle ($sPath, $sMedia) {
-		if (preg_match('#http[s]?://#', $sPath) == 0) {
+		if (preg_match('#http[s]?://#', $sPath) == 0 && is_file($sPath)) {
 			echo '<style type="text/css" media="' . $sMedia . '" >' . "\n" . file_get_contents($sPath) . "\n" . '</style>' . "\n";
 		} else {
 			echo  '<link rel="stylesheet" href="' . $sPath . '"' . ($sMedia ? ' media="'. $sMedia.'"' : '') . " />\n";
